Improvements and Bugfixes

+++ BlueControls/BlueControls/RectangleM.cs
@@ -145,7 +145,7 @@ namespace BlueControls {
 
 
         public RectangleF ZoomAndMoveRect(decimal cZoom, decimal shiftX, decimal shiftY) {
-            return new RectangleF((int)(X * cZoom - shiftX + cZoom / 2), (int)(Y * cZoom - shiftY + cZoom / 2), (int)(Width * cZoom), (int)(Height * cZoom));
+            return new RectangleF((float)(X * cZoom - shiftX + cZoom / 2), (float)(Y * cZoom - shiftY + cZoom / 2), (float)(Width * cZoom), (float)(Height * cZoom));
         }
 
         /// <summary>
